+++ bin/.gitignore
@@ -1,2 +1,1 @@
-/zaposleni/
-/dodavanje/
+/aplikacija/
