+++ src/aplikacija/model/Cetkica.java
