+++ src/aplikacija/model/Softver.java
